% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umf_solve.R
\name{umf_solve}
\alias{umf_solve}
\title{Solves the system of linear equations using UMFPACK}
\usage{
umf_solve(a, b, umf_control = list())
}
\arguments{
\item{a}{an object of class  \code{dgCMatrix}
(see \code{\link[Matrix]{dgCMatrix-class}})}

\item{b}{the vector \eqn{b}}

\item{umf_control}{A named list with control parameters passed to UMFPACK.
See Details.}
}
\value{
the solution \eqn{x}
}
\description{
This function solved the linear equations of the form \eqn{A x = b} using
UMFPACK.
}
\details{
\subsection{UMFPACK control options}{
Argument \code{umf_control} can be used to specify UMFPACK control parameters.
Consult the documentation of UMFPACK for a description of the various control
paramters. \code{umf_control} should be a named list; the names are the
names of the UMFPACK control parameters excluding the prefix \code{UMFPACK}.
With a few exceptions described below, the values are numerical values.
For example,\preformatted{list(SYM_PIVOT_TOLERANCE = 0.01, AGGRESIVE = 1)
}

The values for the control options \code{STRATEGY}, \code{ORDERING} and \code{SCALE}
should be character. In UMFPACK the values of these parameters
should be specified with named numerical constants. For example,
for \code{UMFPACK_STRATEGY} allowed values are the constants
\code{UMFPACK_STRATEGY_AUTO}, \code{UMFPACK_STRATEGY_UNSYMMETRIC} and
\code{UMFPACK_STRATEGY_SYMMETRIC}. In package \code{umfpackr} the values should be
the name of the corresponding constant , again excluding the prefix \code{UMFACK_}.
Exanple:\preformatted{list(STRATEGY = "STRATEGY_UNSYMMETRIC",
     ORDERING = "ORDERING_METIS",
     SCALE    = "SCALE_NONE")
}

}
}
\references{
Dennis, J.E. Jr and Schnabel, R.B. (1997), \emph{Numerical Methods for Unconstrained Optimization
and Nonlinear Equations}, Siam.

Davis, T.A. (2004). A column pre-ordering strategy for the unsymmetric-pattern
multifrontal method. \emph{ACM Trans. Math. Softw.}, \bold{30(2)}, 165–195.

Davis, T.A (2004). Algorithm 832: UMFPACK, an unsymmetric-pattern multifrontal
method. \emph{ACM Trans. Math. Softw.}, \bold{30(2)}, 196–199.

Davis, T.A and Duff, I.S. (1997). An unsymmetric-pattern multifrontal method for
sparse LU factorization. \emph{SIAM J. Matrix Anal. Applic.}, \bold{18(1)}, 140–158.

Davis, T.A  and Duff, I.S (1999). A combined unifrontal/multifrontal method for
unsymmetric sparse matrices. \emph{ACM Trans. Math. Softw.}, \bold{25(1)}, 1–19..
}
