% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umf_solve.R
\name{umf_solve}
\alias{umf_solve}
\title{Solves the system of linear equations \eqn{A x = b} using UMFPACK}
\usage{
umf_solve(a, b)
}
\arguments{
\item{a}{an object of class  \code{dgCMatrix}
(see \code{\link[Matrix]{dgCMatrix-class}})}

\item{b}{the vector \eqn{b}}
}
\value{
the solution \eqn{x}
}
\description{
Solves the system of linear equations \eqn{A x = b} using UMFPACK
}
