% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umf_solve.R
\name{umf_solve}
\alias{umf_solve}
\title{Solves the system of linear equations using UMFPACK}
\usage{
umf_solve(a, b, umf_control = list())
}
\arguments{
\item{a}{an object of class  \code{dgCMatrix}
(see \code{\link[Matrix]{dgCMatrix-class}})}

\item{b}{the vector \eqn{b}}

\item{umf_control}{A named list with control parameters passed to UMFPACK.
Currently only a single control parameter can specified: \code{ordering}, which
specifies the ordering method. Allowed values are \code{"AMD"} (the default),
\code{"CHOLMOD"}, \code{"METIS"} and \code{"BEST"}. See the Vignette
\href{../doc/UMFPACK_interface.pdf}{\emph{UMFPACK interface for R}} for more
details. For example, to use METIS ordering specify
\code{umf_control = list(ORDERING = "METIS")}. The METIS ordering method can
handle larger matrices than the standard AMD method.}
}
\value{
the solution \eqn{x}
}
\description{
This function solved the linear equations of the form \eqn{A x = b} using
UMFPACK.
}
\references{
Dennis, J.E. Jr and Schnabel, R.B. (1997), \emph{Numerical Methods for Unconstrained Optimization
and Nonlinear Equations}, Siam.

Davis, T.A. (2004). A column pre-ordering strategy for the unsymmetric-pattern
multifrontal method. \emph{ACM Trans. Math. Softw.}, \bold{30(2)}, 165–195.

Davis, T.A (2004). Algorithm 832: UMFPACK, an unsymmetric-pattern multifrontal
method. \emph{ACM Trans. Math. Softw.}, \bold{30(2)}, 196–199.

Davis, T.A and Duff, I.S. (1997). An unsymmetric-pattern multifrontal method for
sparse LU factorization. \emph{SIAM J. Matrix Anal. Applic.}, \bold{18(1)}, 140–158.

Davis, T.A  and Duff, I.S (1999). A combined unifrontal/multifrontal method for
unsymmetric sparse matrices. \emph{ACM Trans. Math. Softw.}, \bold{25(1)}, 1–19..
}
\seealso{
\code{\link{umf_solve_nl}}.
}
