% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{umf_solve}
\alias{umf_solve}
\title{Solves the system of linear equations \eqn{A x = b} using UMFPACK}
\usage{
umf_solve(a, b)
}
\arguments{
\item{a}{a \code{dgCMatrix}}

\item{b}{the vector \eqn{b}}
}
\value{
x the solution \eqn{x}
}
\description{
Solves the system of linear equations \eqn{A x = b} using UMFPACK
}

