% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{umf_solve}
\alias{umf_solve}
\title{Solve linear equation using UMFPACK}
\usage{
umf_solve(a, b)
}
\arguments{
\item{b}{the right hand side of A x = b}

\item{A}{a \code{dgCMatrix}}
}
\value{
x the solution x.
}
\description{
Solve linear equation using UMFPACK
}

